SPM

Documentation for SPM, a package for Statistical Process Monitoring.

Package features

Standard control charts

Univariate Shewhart, EWMA, AEWMA, and CUSUM control charts;

MEWMA and MAEWMA control charts;

Control limits

Classical one-sided and two-sided control limits;

Support for multi-chart combinations;

Dynamic control limits based on bootstrap and permutation methods;

Optimization methods

State-of-the-art methods for estimating control limits;

Optimization of control chart parameters against user-defined out-of-control scenarios;

Extensibility to user-made control statistics

Users only need to define the behaviour of the control statistic (struct and sequential update function), everything else is taken care of by the package.

SPM.apply_chart!

SPM.calculate_limit_gradient

SPM.get_limit

SPM.get_limit_value

SPM.get_maxrl

SPM.get_nominal

SPM.get_param

SPM.get_phase1

SPM.get_statistic

SPM.get_t

SPM.get_value

SPM.is_IC

SPM.new_data

SPM.new_data

SPM.run_sim

SPM.saCL!

SPM.set_limit!

SPM.set_nominal!

SPM.set_parameter!

SPM.set_phase1!

SPM.set_statistic!

SPM.update_chart!

SPM.apply_chart! — Method

apply_chart(CH::AbstractChart, x::AbstractVecOrMat)
apply_chart!(CH::AbstractChart, x::AbstractVector)
apply_chart!(CH::AbstractChart, x::AbstractMatrix)

Apply a control chart to a data vector or data matrix x.

SPM.calculate_limit_gradient — Method

calculate_limit_gradient(CH::AbstractChart, rl::Real)

Calculate the gradient for the optimization of the control limit, according to Equation (9) of Capizzi, G., and Masarotto, G. (2016). "Efficient Control Chart Calibration by Simulated Stochastic Approximation". IIE Transactions 48 (1). https://doi.org/10.1080/0740817X.2015.1055392.

SPM.get_limit — Method

get_limit(CH::AbstractChart)

Get the control limit of a control chart.

SPM.get_limit_value — Method

get_limit_value(CH::AbstractChart)

Get the control limit value of a control chart.

SPM.get_maxrl — Method

get_maxrl(CH::AbstractChart)

Get the maximum run length of the control chart.

SPM.get_nominal — Method

get_nominal(CH::AbstractChart)
get_nominal_value(CH::AbstractChart)

Get the nominal properties of a control chart.

SPM.get_param — Method

get_param(CH::AbstractChart)
set_param!(CH::AbstractChart, par)

Get and set the parameters of the control chart statistic.

SPM.get_phase1 — Method

get_phase1(CH::AbstractChart)

Get the Phase 1 information of a control chart.

SPM.get_statistic — Method

get_statistic(CH::AbstractChart)

Get the statistic of a control chart.

SPM.get_t — Method

get_t(CH::AbstractChart)

Get the current time point from a control chart.

SPM.get_value — Method

get_value(CH::AbstractChart)

Get the current value of the control chart statistic.

SPM.is_IC — Method

is_IC(CH::AbstractChart)
is_OC(CH::AbstractChart)

Check whether the control chart is in control or out of control.

SPM.new_data — Method

new_data(CH::AbstractChart)

Simulate a new observation for the control chart.

SPM.new_data — Method

new_data(P1::Phase1Data)

Generates a new observation via bootstrap, based on the observed Phase I data.

SPM.run_sim — Method

run_sim(CH::AbstractChart)

Simulates a run length for the control chart CH by sampling new data from the Phase I object.

SPM.saCL! — Method

saCL(CH::ControlChart, rlsim::Function, kw...)

Computes the control limit for a control chart such that it satisfies E[RL] = Arl0.

Keyword arguments are:

SPM.set_limit! — Method

function set_limit!(CH::AbstractChart, limit::AbstractLimit)

Set the control limit of a control chart.

SPM.set_nominal! — Method

set_nominal!(CH::AbstractChart, nominal::NominalProperties)

Set the nominal properties of a control chart.

SPM.set_parameter! — Method

function set_parameter!(CH::C, par)

Set the parameters of the control chart statistic.

SPM.set_phase1! — Method

set_phase1!(CH::AbstractChart, phase1::AbstractPhase1)

Set the Phase 1 information of a control chart.

SPM.set_statistic! — Method

function set_statistic!(CH::AbstractChart, statistic::AbstractStatistic)

Set the statistic of a control chart.

SPM.update_chart! — Method

update_chart!(CH::AbstractChart, x)

Update the control chart using a new observation x.