Control charts

SPM.get_design — Method

get_design(CH::AbstractChart)

Get the designs of the control chart statistic.

SPM.get_limit — Method

get_limit(CH::AbstractChart)

Get the control limit of a control chart.

SPM.get_limit_value — Method

get_limit_value(CH::AbstractChart)

Get the control limit value of a control chart.

SPM.get_maxrl — Method

get_maxrl(CH::AbstractChart)

Get the maximum run length of the control chart.

SPM.get_nominal — Method

get_nominal(CH::AbstractChart)

Get the nominal properties of a control chart.

SPM.get_nominal_value — Method

get_nominal(CH::AbstractChart)

Get the nominal value of a control chart.

SPM.get_phase2 — Method

get_phase2(CH::AbstractChart)

Get the phase 2 information of a control chart.

SPM.get_statistic — Method

get_statistic(CH::AbstractChart)

Get the statistic of a control chart.

SPM.get_t — Method

get_t(CH::AbstractChart)

Get the current time point of a control chart.

SPM.get_value — Method

get_value(CH::AbstractChart)

Get the current value of the control chart statistic.

SPM.is_IC — Method

is_IC(CH::AbstractChart)
is_OC(CH::AbstractChart)

Check whether the control chart is in control or out of control.

SPM.is_IC_vec — Method

is_IC_vec(CH::MultipleControlChart)
is_OC_vec(CH::MultipleControlChart)

Check whether each individual control chart that makes up a multiple control chart is in control or out of control.

Returns

A vector of Bool, whose length is equal to the number of individual statistics.

SPM.new_data! — Method

new_data(CH::AbstractChart)

Simulate a new observation for the control chart from the phase 2 data, eventually modifying the underlying phase 2 object.

SPM.new_data — Method

new_data(CH::AbstractChart)

Simulate a new observation based on the control chart's Phase II object.

SPM.set_design! — Method

get_design(CH::AbstractChart)

Set the designs of the control chart statistic.

SPM.set_limit! — Method

function set_limit!(CH::AbstractChart, limit::AbstractLimit)
function set_limit!(CH::AbstractChart, h::Float64)
function set_limit!(CH::MultipleControlChart, h::Vector{Float64})
function set_limit!(CH::MultipleControlChart, h::Float64)
function set_limit!(CH::MultipleControlChart, h::Float64, j::Int)

Set the control limit of a control chart.

Returns

The new control limit.

SPM.set_nominal! — Method

set_nominal!(CH::AbstractChart, nominal::NominalProperties)

Set the nominal properties of a control chart.

SPM.set_phase2! — Method

set_phase2!(CH::AbstractChart, phase2::AbstractPhase2)

Set the Phase 2 information of a control chart to simulate run lenghts.

SPM.set_statistic! — Method

function set_statistic!(CH::AbstractChart, statistic::AbstractStatistic)

Set the statistic of a control chart.

Returns

The new value of the statistic.

SPM.set_t! — Method

function set_t!(CH::AbstractChart, t)

Set the current time value of a control chart.

Returns

The new time value of the control chart.

SPM.set_value! — Method

function set_value!(CH::AbstractChart, value)

Set the value of the statistic of a control chart.

Returns

The new value of the control chart's statistic.

SPM.shallow_copy_sim — Method

shallow_copy_sim(CH::ControlChart)

Create a shallow copy of a control chart, so that only the statistic and the control limit are copied. This is done to prevent copying eventual Phase 2 data multiple times and thus reduce computational effort when optimizing the control limit and the chart tuning parameters.

SPM.update_chart! — Method

update_chart!(CH::AbstractChart, x)

Update the control chart inplace using a new observation x.

SPM.update_chart — Method

update_chart(CH::AbstractChart, x)

Update the control chart without modifying it using a new observation x.

SPM.update_limit! — Method

update_limit!(CH::AbstractChart, x)

Update the dynamic control limit of a control chart inplace.

SPM.run_path_sim — Method

run_path_sim(CH::AbstractChart; maxiter)
run_path_sim(CH::MultipleControlChart; maxiter)

Simulates a run length path for the control chart CH by sampling new data from its Phase II object.

Inputs

CH::AbstractChart - A control chart.

maxiter::Real - The maximum value of the run length. Defaults to min(maxrl(CH), 10*get_nominal_value(CH))

Returns

A vector containing the simulated values of the control chart.

SPM.run_sim — Method

run_sim(CH::AbstractChart, DGP::AbstractPhase2)

Simulates a run length for the control chart CH by sampling new data from the provided data-generating process DGP.

Inputs

CH - A control chart.

DGP - An AbstractPhase2 object.

Returns

An Int representing the simulated run length.

SPM.run_sim — Method

run_sim(CH::AbstractChart)

Simulates a run length for the control chart CH by sampling new data from its Phase II object.

Inputs

CH - A control chart.

Returns

An Int representing the simulated run length.

SPM.run_sim_oc — Method

run_sim_oc(CH::AbstractChart; shift = 0.0)

Simulates a run length under location shift for the control chart CH by sampling new data from its Phase II object.

Inputs

CH::AbstractChart: A control chart.

shift::Float64 - The magnitude of location shift.

Returns

An Int representing the simulated run length.

SPM.run_sim_sa — Method

run_sim_sa(CH::AbstractChart, maxiter::Real, delta::Real)

Simulates a run length for the control chart CH by sampling new data from the Phase II object, to be used by the stochastic approximation algorithm implemented in the saCL! function.

Inputs

CH - A control chart.

maxiter - The maximum value of the run length.

delta - A value controlling how much the control limit must be shifted for the gain estimation during the first stage.

Returns

A NamedTuple containing the simulated run length, rl, the simulated run length with control limit shifted by delta, rlPlus, and the simulated run length with control limit shifted by -delta, rlMinus.

SPM.apply_chart! — Method

apply_chart(CH::AbstractChart, x::AbstractVecOrMat)
apply_chart!(CH::AbstractChart, x::AbstractVector)
apply_chart!(CH::AbstractChart, x::AbstractMatrix)

Apply a control chart to a data vector or data matrix x.