Control limits

StatisticalProcessMonitoring.OneSidedFixedLimit — Type

OneSidedFixedLimit(h::Float64, upw::Bool)

Classical fixed one-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit.

if upw == true, $RL = \inf\{t : C_t > h\}$

if upw == false, $RL = \inf\{t : C_t < -h\}$

Note that by definition, h > 0.

StatisticalProcessMonitoring.TwoSidedFixedLimit — Type

TwoSidedFixedLimit(h::Float64)

Classical fixed two-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit:

$RL = \inf\{t > 0 : |C_t| > h\}$.

Note that by definition, h > 0.

StatisticalProcessMonitoring.OneSidedBootstrapLimit — Type

OneSidedBootstrapLimit{T} <: BootstrapLimit{T}

A one-sided bootstrap limit with constant false-alarm rate.

Fields

sim::Vector{T}: The vector of simulated statistics.

h::T: The current value of the control limit.

upw::Bool: Whether the control limit is an upper or lower control limit.

Constructors

OneSidedBootstrapLimit(S::AbstractStatistic, upw, B::Int): Create a new OneSidedBootstrapLimit object. The argument S is an AbstractStatistic object. The argument upw determines whether the bootstrap is one-sided and upper-tailed or lower-tailed. The argument B is an integer indicating the number of bootstrap replications.

StatisticalProcessMonitoring.OneSidedCurvedLimit — Type

OneSidedCurvedLimit(h::Float64, upw::Bool)
OneSidedCurvedLimit(h::Vector{T}, upw::Vector{Bool})

Curved one-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit.

if upw == true, $RL = \inf\{t : C_t > h\cdot f(t)\}$

if upw == false, $RL = \inf\{t : C_t < -h\cdot f(t)\}$

Note that by definition, h > 0.

StatisticalProcessMonitoring.TwoSidedBootstrapLimit — Type

TwoSidedBootstrapLimit{T} <: BootstrapLimit{T}

A two-sided bootstrap limit with constant false-alarm rate.

Fields

sim::Vector{T}: The vector of simulated statistics.

h::Vector{T}: The current value of the control limits.

Constructors

TwoSidedBootstrapLimit(S::AbstractStatistic, B::Int): Create a new TwoSidedBootstrapLimit object. The argument S is an AbstractStatistic object. The argument B is an integer indicating the number of bootstrap replications.

StatisticalProcessMonitoring.TwoSidedCurvedLimit — Type

TwoSidedCurvedLimit(h::Float64)
TwoSidedCurvedLimit(h::Vector{T})

Curved one-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit.

$RL = \inf\{t > 0 : |C_t| > h\cdot f(t)\}$.

Note that by definition, h > 0.

StatisticalProcessMonitoring.update_value! — Method

update_value!(L::BootstrapLimit, NM::ARL)
update_value!(L::BootstrapLimit, NM::QRL)

Update the value of the BootstrapLimit object using the nominal value obtained from the ARL or QRL object.

Arguments

L::BootstrapLimit: The BootstrapLimit object to be updated.

NM: The ARL or QRL object used to obtain the value for the update.