SPM

Documentation for SPM, a package for Statistical Process Monitoring.

Package features

Standard control charts

Univariate Shewhart, EWMA, AEWMA, and CUSUM control charts;

MEWMA and MAEWMA control charts;

Control limits

Classical one-sided and two-sided control limits;

Support for multi-chart combinations;

Dynamic control limits based on bootstrap and permutation methods;

Optimization methods

State-of-the-art methods for estimating control limits;

Optimization of control chart parameters against user-defined out-of-control scenarios;

Extensibility to user-made control statistics

Users only need to define the behaviour of the control statistic (struct and sequential update function), everything else is taken care of by the package.

SPM.ARL

SPM.CUSUM

SPM.EWMA

SPM.OneSidedCurvedLimit

SPM.OneSidedFixedLimit

SPM.QRL

SPM.TwoSidedCurvedLimit

SPM.TwoSidedFixedLimit

SPM.apply_chart!

SPM.bisectionCL

SPM.bisectionCL!

SPM.calculate_limit_gradient

SPM.combinedCL

SPM.combinedCL!

SPM.get_limit

SPM.get_limit_value

SPM.get_maxrl

SPM.get_nominal

SPM.get_nominal_value

SPM.get_parameter

SPM.get_phase1

SPM.get_statistic

SPM.get_t

SPM.get_value

SPM.is_IC

SPM.is_IC_vec

SPM.new_data

SPM.new_data!

SPM.run_sim

SPM.run_sim

SPM.run_sim_oc

SPM.run_sim_sa

SPM.saCL

SPM.saCL!

SPM.set_limit!

SPM.set_nominal!

SPM.set_parameter!

SPM.set_phase1!

SPM.set_statistic!

SPM.shallow_copy_sim

SPM.update_chart

SPM.update_chart!

SPM.update_limit!

SPM.update_value!

SPM.apply_chart! — Method

apply_chart(CH::AbstractChart, x::AbstractVecOrMat)
apply_chart!(CH::AbstractChart, x::AbstractVector)
apply_chart!(CH::AbstractChart, x::AbstractMatrix)

Apply a control chart to a data vector or data matrix x.

SPM.bisectionCL! — Method

bisectionCL!(CH::ControlChart; hmax, kw...)

Computes the control limit to satisfy the nominal properties of a control chart, using the bisection algorithm (see for instance Qiu, 2013)

Inputs

CH - A control chart.

kw... - Keyword arguments that control the behaviour of the algorithm.

Keyword arguments:

rlsim - A function that generates new data with signature rlsim(CH; maxiter). If left unspecified, defaults to run_sim.

hmin - The minimum value of the control limit, defaults to sqrt(eps()).

hmax - The maximum value for the control limit.

maxiter - The maximum number of bisection iterations.

nsims - The number of run lengths used to estimate the target nominal property.

trunc - The maximum run length after which it is truncated, to avoid excessive computations.

x_tol - Absolute tolerance for the algorithm, which is ended if $h^{(k+1)} - h^{(k)} < x_{\text{tol}}$

f_tol - Absolute tolerance for the algorithm, which is ended if $\text{target}(h^{(k+1)}) - \text{target}(h^{(k)}) < f_{\text{tol}}$

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

SPM.bisectionCL — Method

bisectionCL(CH::ControlChart; kw...)

Applies the bisection algorithm to find the control limit of a control chart without modifying the control chart object CH. See the documentation of bisectionCL! for more information about the algorithm and keyword arguments.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

SPM.calculate_limit_gradient — Method

calculate_limit_gradient(CH::AbstractChart, rl::Real)
calculate_limit_gradient(nominal::ARL, rl)
calculate_limit_gradient(nominal::QRL, rl)

Calculate the gradient for the optimization of the control limit.

If the control chart nominal attribute is of type ARL, then the gradient is calculated according to Equation (9) of Capizzi and Masarotto (2016).

If the control chart nominal attribute is of type QRL, then the gradient is calculated using the recursion on page 280 of Capizzi and Masarotto (2009)

References

Capizzi, G., & Masarotto, G. (2016). "Efficient Control Chart Calibration by Simulated Stochastic Approximation". IIE Transactions 48 (1). https://doi.org/10.1080/0740817X.2015.1055392.

Capizzi, G. & Masarotto, G. (2009) Bootstrap-based design of residual control charts, IIE Transactions, 41:4, 275-286, DOI: https://doi.org/10.1080/07408170802120059

SPM.combinedCL! — Method

combinedCL!(CH::ControlChart; kw...)

Computes the control limit to satisfy the nominal properties of a control chart, using the bisection algorithm (see for instance Qiu, 2013). The control limit upper bound hmax for the bisection algorithm is found using the stochastic approximation algorithm of Capizzi and Masarotto (2016)

Inputs

CH - A control chart.

kw... - Keyword arguments that control the behaviour of the algorithm.

Keyword arguments:

rlsim - A function that generates new data with signature rlsim(CH; maxiter). If left unspecified, defaults to run_sim.

hmin - The minimum value of the control limit, defaults to sqrt(eps()).

hmax - The maximum value for the control limit.

maxiter - The maximum number of bisection iterations.

nsims - The number of run lengths used to estimate the target nominal property.

trunc - The maximum run length after which it is truncated, to avoid excessive computations.

x_tol - Absolute tolerance for the algorithm, which is ended if $h^{(k+1)} - h^{(k)} < x_{\text{tol}}$

f_tol - Absolute tolerance for the algorithm, which is ended if $\text{target}(h^{(k+1)}) - \text{target}(h^{(k)}) < f_{\text{tol}}$

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

Capizzi, G., & Masarotto, G. (2016). Efficient control chart calibration by simulated stochastic approximation. IIE Transactions, 48(1), 57-65. https://doi.org/10.1080/0740817X.2015.1055392

SPM.combinedCL — Method

combinedCL(CH::ControlChart; kw...)

Applies the bisection algorithm to find the control limit of a control chart without modifying the control chart object CH. The control limit upper bound hmax for the bisection algorithm is found using the stochastic approximation algorithm of Capizzi and Masarotto (2016). See the documentation of combinedCL! for more information about the algorithm and keyword arguments.

Keyword arguments:

See the documentation of combinedCL! for a list of keyword arguments.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

Capizzi, G., & Masarotto, G. (2016). Efficient control chart calibration by simulated stochastic approximation. IIE Transactions, 48(1), 57-65. https://doi.org/10.1080/0740817X.2015.1055392

SPM.get_limit — Method

get_limit(CH::AbstractChart)

Get the control limit of a control chart.

SPM.get_limit_value — Method

get_limit_value(CH::AbstractChart)

Get the control limit value of a control chart.

SPM.get_maxrl — Method

get_maxrl(CH::AbstractChart)

Get the maximum run length of the control chart.

SPM.get_nominal — Method

get_nominal(CH::AbstractChart)

Get the nominal properties of a control chart.

SPM.get_nominal_value — Method

get_nominal(CH::AbstractChart)

Get the nominal value of a control chart.

SPM.get_parameter — Method

get_parameter(CH::AbstractChart)

Get the parameters of the control chart statistic.

SPM.get_phase1 — Method

get_phase1(CH::AbstractChart)

Get the Phase 1 information of a control chart.

SPM.get_statistic — Method

get_statistic(CH::AbstractChart)

Get the statistic of a control chart.

SPM.get_t — Method

get_t(CH::AbstractChart)

Get the current time point of a control chart.

SPM.get_value — Method

get_value(CH::AbstractChart)

Get the current value of the control chart statistic.

SPM.is_IC — Method

is_IC(CH::AbstractChart)
is_OC(CH::AbstractChart)

Check whether the control chart is in control or out of control.

SPM.is_IC_vec — Method

is_IC_vec(CH::MultipleControlChart)
is_OC_vec(CH::MultipleControlChart)

Check whether each individual control chart that makes up a multiple control chart is in control or out of control.

Returns

A vector of Bool, whose length is the number of individual statstics.

SPM.new_data! — Method

new_data(CH::AbstractChart)

Simulate a new observation for the control chart from the Phase 1 data, eventually modifying the underlying phase 1 object.

SPM.new_data — Method

new_data(CH::AbstractChart)

Simulate a new observation for the control chart from the Phase 1 data.

SPM.run_sim — Method

run_sim(CH::AbstractChart, DGP::AbstractPhase1)

Simulates a run length for the control chart CH by sampling new data from the provided data-generating process DGP.

Inputs

CH - A control chart.

DGP - An AbstractPhase1 object.

Returns

An Int.

SPM.run_sim — Method

run_sim(CH::AbstractChart)

Simulates a run length for the control chart CH by sampling new data from its Phase I object.

Inputs

CH - A control chart.

Returns

An Int.

SPM.run_sim_oc — Method

run_sim_oc(CH::AbstractChart; shift = 0.0)

Simulates a run length under location shift for the control chart CH by sampling new data from its Phase I object.

Inputs

CH - A control chart.

shift - A location shift.

Returns

An Int.

SPM.run_sim_sa — Method

run_sim_sa(CH::AbstractChart, maxiter::Real, deltaSA::Real)

Simulates a run length for the control chart CH by sampling new data from the Phase I object, to be used by the stochastic approximation algorithm implemented in the saCL! function.

Inputs

CH - A control chart.

maxiter - The maximum value of the run length.

deltaSA - A value controlling how much the control limit must be shifted for the gain estimation during the first stage.

Returns

A NamedTuple containing the simulated run length, rl, the simulated run length with control limit shifted by deltaSA, rlPlus, and the simulated run length with control limit shifted by -deltaSA, rlMinus.

SPM.saCL! — Method

saCL!(CH::ControlChart; kw...)

Computes the control limit to satisfy the nominal properties of a control chart, using the stochastic approximation algorithm described in Capizzi and Masarotto (2016).

Inputs

CH - A control chart.

kw... - Keyword arguments that control the behaviour of the algorithm. For more information about the specifics of each keyword argument, see Capizzi and Masarotto (2016).

Keyword arguments:

rlsim - A function that generates new data with signature rlsim(CH; maxiter, deltaSA). If left unspecified, defaults to run_sim_sa. See the help for run_sim_sa for more information.

Nfixed - The number of iterations for the gain estimation stage.

Afixed - The fixed gain during the gain estimation stage.

Amin - The minimum allowed value of gain.

Amax - The maximum allowed value of gain.

deltaSA - The shift in control limit used during the gain estimation stage.

q - The power that controls the denominator in the Robbins-Monro algorithm.

gamma - The precision parameter for the stopping criterion of the algorithm.

Nmin - The minimum number of iterations required for the algorithm to end.

z - The quantile of the Normal(0,1) that controls the probability of the stopping criterion being satisfied.

Cmrl - The inflation factor for the maximum number of iterations the run length may run for.

maxiter - Maximum number of iterations before the algorithm is forcibly ended.

verbose - Whether to print information to the user about the state of the optimization.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Capizzi, G., & Masarotto, G. (2016). "Efficient Control Chart Calibration by Simulated Stochastic Approximation". IIE Transactions 48 (1). https://doi.org/10.1080/0740817X.2015.1055392.

SPM.saCL — Method

saCL(CH::ControlChart; kw...)

Applies the stochastic approximation algorithm of Capizzi and Masarotto (2016) without modifying the control chart object CH. See the documentation of saCL! for more information about the algorithm and keyword arguments.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Capizzi, G., & Masarotto, G. (2016). "Efficient Control Chart Calibration by Simulated Stochastic Approximation". IIE Transactions 48 (1). https://doi.org/10.1080/0740817X.2015.1055392.

SPM.set_limit! — Method

function set_limit!(CH::AbstractChart, limit::AbstractLimit)

Set the control limit of a control chart.

SPM.set_nominal! — Method

set_nominal!(CH::AbstractChart, nominal::NominalProperties)

Set the nominal properties of a control chart.

SPM.set_parameter! — Method

get_parameter(CH::AbstractChart)

Set the parameters of the control chart statistic.

SPM.set_phase1! — Method

set_phase1!(CH::AbstractChart, phase1::AbstractPhase1)

Set the Phase 1 information of a control chart.

SPM.set_statistic! — Method

function set_statistic!(CH::AbstractChart, statistic::AbstractStatistic)

Set the statistic of a control chart.

SPM.shallow_copy_sim — Method

shallow_copy_sim(CH::AbstractChart)

Create a shallow copy of a control chart, so that only the statistic and the control limit are copied. This is done to prevent copying the Phase 1 data multiple times and thus reduce computational effort.

SPM.update_chart! — Method

update_chart!(CH::AbstractChart, x)

Update the control chart inplace using a new observation x.

SPM.update_chart — Method

update_chart(CH::AbstractChart, x)

Update the control chart without modifying it using a new observation x.

SPM.update_limit! — Method

update_limit!(CH::AbstractChart, x)

Update the dynamic control limit of a control chart inplace.

SPM.update_value! — Method

get_value(L::BootstrapLimit, NM::ARL)
get_value(L::BootstrapLimit, NM::QRL)

SPM.ARL — Type

ARL(value)

Value of the in-control average run length of the control chart, i.e. if $RL = \inf{t > 0: \text{Chart detects OC}}$ is the run length, then the average run length $ARL$ is

$ARL = \mathbb{E}[RL|\tau = +\infty]$,

where $\{\tau = +\infty\}$ represents the process being in-control.

SPM.CUSUM — Type

CUSUM(k, value, upw::Bool)

CUSUM statistic with parameter k and initial value value.

The update mechanism based on a new observation x is given by:

if upw == true, then $value = \max\{0, value + x - k\}$;

if upw == false, then $value = \min\{0, value + x + k\}$.

References

Page, E. S. (1954). Continuous Inspection Schemes. Biometrika, 41(1/2), 100. https://doi.org/10.2307/2333009

SPM.EWMA — Type

EWMA(λ, value)

Exponentially weighted moving average with parameter λ and initial value value.

The update mechanism based on a new observation x is given by

$value = (1-λ)*value + λ * x$.

References

Roberts, S. W. (1959). Control Chart Tests Based on Geometric Moving Averages. Technometrics, 1(3), 239-250. https://doi.org/10.1080/00401706.1959.10489860

SPM.OneSidedCurvedLimit — Type

OneSidedCurvedLimit(h::Float64, upw::Bool)
OneSidedCurvedLimit(h::Vector{T}, upw::Vector{Bool})

Curved one-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit.

if upw == true, $RL = \inf\{t : C_t > h\cdot f(t)\}$

if upw == false, $RL = \inf\{t : C_t < -h\cdot f(t)\}$

Note that h > 0 by the way it is defined.

SPM.OneSidedFixedLimit — Type

OneSidedFixedLimit(h::Float64, upw::Bool)

Classical fixed one-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit.

if upw == true, $RL = \inf\{t : C_t > h\}$

if upw == false, $RL = \inf\{t : C_t < h\}$

Note that h > 0 by the way it is defined.

SPM.QRL — Type

QRL(value, qtl)

Value of the in-control quantile of the run length of the control chart, i.e. if $RL = \inf{t > 0 : \text{Chart detects OC}}$ is the run length, then value is the value of the qtl-level quantile of the distribution of $RL$ if the process is in-control.

SPM.TwoSidedCurvedLimit — Type

TwoSidedCurvedLimit(h::Float64)
TwoSidedCurvedLimit(h::Vector{T})

Curved one-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit.

$RL = \inf\{t > 0 : |C_t| > h\cdot f(t)\}$.

Note that h > 0 by the way it is defined.

SPM.TwoSidedFixedLimit — Type

TwoSidedFixedLimit(h::Float64)

Classical fixed two-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit:

$RL = \inf\{t > 0 : |C_t| > h\}$.

Note that h > 0 by the way it is defined.