SPM

Documentation for SPM, a package for Statistical Process Monitoring.

Package features

Standard control charts

Univariate Shewhart, EWMA, AEWMA, and CUSUM control charts;

MEWMA and MAEWMA control charts;

Control limits

Classical one-sided and two-sided control limits;

Support for multi-chart combinations;

Dynamic control limits based on bootstrap and permutation methods;

Optimization methods

State-of-the-art methods for estimating control limits;

Optimization of control chart parameters against user-defined out-of-control scenarios;

Extensibility to user-made control statistics

Users only need to define the behaviour of the control statistic (struct and sequential update function), everything else is taken care of by the package.

SPM.AEWMA

SPM.ARL

SPM.BlockBootstrap

SPM.CUSUM

SPM.EWMA

SPM.LocationScaleStatistic

SPM.MEWMA

SPM.OneSidedCurvedLimit

SPM.OneSidedFixedLimit

SPM.Phase2

SPM.Phase2Distribution

SPM.QRL

SPM.StationaryBootstrap

SPM.TwoSidedCurvedLimit

SPM.TwoSidedFixedLimit

SPM.apply_chart!

SPM.bisectionCL

SPM.bisectionCL!

SPM.calculate_limit_gradient

SPM.combinedCL

SPM.combinedCL!

SPM.doubleBootstrap

SPM.doubleBootstrap!

SPM.get_design

SPM.get_limit

SPM.get_limit_value

SPM.get_maxrl

SPM.get_nominal

SPM.get_nominal_value

SPM.get_phase1

SPM.get_statistic

SPM.get_t

SPM.get_value

SPM.get_value

SPM.is_IC

SPM.is_IC_vec

SPM.new_data

SPM.new_data

SPM.new_data!

SPM.optimize_design

SPM.optimize_design!

SPM.optimize_grid

SPM.optimize_limit

SPM.optimize_limit!

SPM.optimize_nlopt

SPM.run_path_sim

SPM.run_sim

SPM.run_sim

SPM.run_sim_oc

SPM.run_sim_sa

SPM.saCL

SPM.saCL!

SPM.set_design!

SPM.set_limit!

SPM.set_nominal!

SPM.set_phase1!

SPM.set_statistic!

SPM.set_value!

SPM.set_value!

SPM.shallow_copy_sim

SPM.update_chart

SPM.update_chart!

SPM.update_limit!

SPM.update_statistic

SPM.update_value!

SPM.apply_chart! — Method

apply_chart(CH::AbstractChart, x::AbstractVecOrMat)
apply_chart!(CH::AbstractChart, x::AbstractVector)
apply_chart!(CH::AbstractChart, x::AbstractMatrix)

Apply a control chart to a data vector or data matrix x.

SPM.bisectionCL! — Method

bisectionCL!(CH::ControlChart[; rlsim::Function, settings::OptSettings])

Computes the control limit to satisfy the nominal properties of a control chart, using the bisection algorithm (see for instance Qiu, 2013)

Inputs

CH - A control chart.

rlsim - A function that generates a run length for the control chart with signature rlsim(CH; maxiter). If left unspecified, defaults to run_sim. See the help for run_sim for more information about the signature of the function.

settings - An OptSettings objects which contains variables that control the behaviour of the algorithm. See the Accepted settings section below for information about the settings that control the behaviour of the algorithm. For more information about the specifics of each keyword argument, see for instance Qiu (2013).

Accepted settings:

hmin_bi - The minimum value of the control limit, defaults to sqrt(eps()).

hmax_bi - The maximum value for the control limit.

maxiter_bi - The maximum number of bisection iterations.

nsims_bi - The number of run lengths used to estimate the target nominal property.

trunc_bi - The maximum run length after which it is trunc_biated, to avoid excessive computations.

x_tol_bi - Absolute tolerance for the algorithm, which is ended if $h^{(k+1)} - h^{(k)} < x_{\text{tol}}$

f_tol_bi - Absolute tolerance for the algorithm, which is ended if $\text{target}(h^{(k+1)}) - \text{target}(h^{(k)}) < f_{\text{tol}}$

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

SPM.bisectionCL — Method

bisectionCL(CH::ControlChart; kw...)

Applies the bisection algorithm to find the control limit of a control chart without modifying the control chart object CH. See the documentation of bisectionCL! for more information about the algorithm and keyword arguments.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

SPM.calculate_limit_gradient — Method

calculate_limit_gradient(CH::AbstractChart, rl::Real)
calculate_limit_gradient(nominal::ARL, rl)
calculate_limit_gradient(nominal::QRL, rl)

Calculate the gradient for the optimization of the control limit.

If the control chart nominal attribute is of type ARL, then the gradient is calculated according to Equation (9) of Capizzi and Masarotto (2016).

If the control chart nominal attribute is of type QRL, then the gradient is calculated using the recursion on page 280 of Capizzi and Masarotto (2009)

References

Capizzi, G., & Masarotto, G. (2016). "Efficient Control Chart Calibration by Simulated Stochastic Approximation". IIE Transactions 48 (1). https://doi.org/10.1080/0740817X.2015.1055392.

Capizzi, G. & Masarotto, G. (2009) Bootstrap-based design of residual control charts, IIE Transactions, 41:4, 275-286, DOI: https://doi.org/10.1080/07408170802120059

SPM.combinedCL! — Method

combinedCL!(CH::ControlChart[; rlsim::Function, settings::OptSettings])

Computes the control limit to satisfy the nominal properties of a control chart, using the bisection algorithm (see for instance Qiu, 2013). The control limit upper bound hmax_bi for the bisection algorithm is found using the stochastic approximation algorithm of Capizzi and Masarotto (2016)

Inputs

CH - A control chart.

rlsim - A function that generates a run length for the control chart with signature rlsim(CH; maxiter). If left unspecified, defaults to run_sim. See the help for run_sim for more information about the signature of the function.

settings - An OptSettings objects which contains variables that control the behaviour of the algorithm. See the Accepted settings section below for information about the settings that control the behaviour of the algorithm. For more information about the specifics of each keyword argument, see for instance Qiu (2013).

Accepted settings

Bisection algorithm

rlsim - A function that generates new data with signature rlsim(CH; maxiter_bi). If left unspecified, defaults to run_sim.

hmin_bi - The minimum value of the control limit, defaults to sqrt(eps()).

hmax_bi - The maximum value for the control limit.

maxiter_bi - The maximum number of bisection iterations.

nsims_bi - The number of run lengths used to estimate the target nominal property.

trunc_bi - The maximum run length after which it is trunc_biated, to avoid excessive computations.

x_tol_bi - Absolute tolerance for the algorithm, which is ended if $h^{(k+1)} - h^{(k)} < x_{\text{tol}}$

f_tol_bi - Absolute tolerance for the algorithm, which is ended if $\text{target}(h^{(k+1)}) - \text{target}(h^{(k)}) < f_{\text{tol}}$

SA algorithm

Nfixed_sa - The number of iterations for the gain estimation stage.

Afixed_sa - The fixed gain during the gain estimation stage.

Amin_sa - The minimum allowed value of gain.

Amax_sa - The maximum allowed value of gain.

deltaSA_sa - The shift in control limit used during the gain estimation stage.

q_sa - The power that controls the denominator in the Robbins-Monro algorithm.

gamma_sa - The precision parameter for the stopping criterion of the algorithm.

Nmin_sa - The minimum number of iterations required for the algorithm to end.

z_sa - The quantile of the Normal(0,1) that controls the probability of the stopping criterion being satisfied.

Cmrl_sa - The inflation factor for the maximum number of iterations the run length may run for.

maxiter_sa - Maximum number of iterations before the algorithm is forcibly ended.

verbose_sa - Whether to print information to the user about the state of the optimization.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

Capizzi, G., & Masarotto, G. (2016). Efficient control chart calibration by simulated stochastic approximation. IIE Transactions, 48(1), 57-65. https://doi.org/10.1080/0740817X.2015.1055392

SPM.combinedCL — Method

combinedCL(CH::ControlChart; kw...)

Applies the bisection algorithm to find the control limit of a control chart without modifying the control chart object CH. The control limit upper bound hmax_bi for the bisection algorithm is found using the stochastic approximation algorithm of Capizzi and Masarotto (2016). See the documentation of combinedCL! for more information about the algorithm and keyword arguments.

Keyword arguments:

See the documentation of combinedCL! for a list of keyword arguments.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

Capizzi, G., & Masarotto, G. (2016). Efficient control chart calibration by simulated stochastic approximation. IIE Transactions, 48(1), 57-65. https://doi.org/10.1080/0740817X.2015.1055392

SPM.doubleBootstrap! — Method

doubleBootstrap!(CH::ControlChart[; rlsim::Function, settings::OptSettings])

Computes the control limit to satisfy the nominal properties of a control chart, using the bisection algorithm on bootstrapped paths (see for instance Qiu, 2013).

Inputs

CH - A control chart.

rlsim - A function that generates a path of the control chart statistic with signature rlsim(CH; maxiter). If left unspecified, defaults to run_path_sim. See the help for run_path_sim for more information about the signature of the function.

settings - An OptSettings objects which contains variables that control the behaviour of the algorithm. See the Accepted settings section below for information about the settings that control the behaviour of the algorithm. For more information about the specifics of each keyword argument, see for instance Qiu (2013).

Accepted settings:

maxiter_bi - The maximum number of bisection iterations.

nsims_bi - The number of run lengths used to estimate the target nominal property.

trunc_bi - The maximum run length after which the run length is truncated, to avoid excessive computations.

x_tol_bi - Absolute tolerance for the algorithm, which is ended if $h^{(k+1)} - h^{(k)} < x_{\text{tol}}$

f_tol_bi - Absolute tolerance for the algorithm, which is ended if $\text{target}(h^{(k+1)}) - \text{target}(h^{(k)}) < f_{\text{tol}}$

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

SPM.doubleBootstrap — Method

doubleBootstrap(CH::ControlChart; kw...)

Applies the bisection algorithm on simulated run length paths to find the control limit of a control chart without modifying the control chart object CH. See the documentation of doubleBootstrap! for more information about the algorithm and keyword arguments.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Qiu, P. (2013). Introduction to Statistical Process Control. CRC Press.

SPM.get_design — Method

get_design(CH::AbstractChart)

Get the designs of the control chart statistic.

SPM.get_limit — Method

get_limit(CH::AbstractChart)

Get the control limit of a control chart.

SPM.get_limit_value — Method

get_limit_value(CH::AbstractChart)

Get the control limit value of a control chart.

SPM.get_maxrl — Method

get_maxrl(CH::AbstractChart)

Get the maximum run length of the control chart.

SPM.get_nominal — Method

get_nominal(CH::AbstractChart)

Get the nominal properties of a control chart.

SPM.get_nominal_value — Method

get_nominal(CH::AbstractChart)

Get the nominal value of a control chart.

SPM.get_phase1 — Method

get_phase1(CH::AbstractChart)

Get the Phase 1 information of a control chart.

SPM.get_statistic — Method

get_statistic(CH::AbstractChart)

Get the statistic of a control chart.

SPM.get_t — Method

get_t(CH::AbstractChart)

Get the current time point of a control chart.

SPM.get_value — Method

get_value(CH::AbstractChart)

Get the current value of the control chart statistic.

SPM.get_value — Method

get_value(stat::AbstractStatistic)

Get the current value of a statistic.

SPM.is_IC — Method

is_IC(CH::AbstractChart)
is_OC(CH::AbstractChart)

Check whether the control chart is in control or out of control.

SPM.is_IC_vec — Method

is_IC_vec(CH::MultipleControlChart)
is_OC_vec(CH::MultipleControlChart)

Check whether each individual control chart that makes up a multiple control chart is in control or out of control.

Returns

A vector of Bool, whose length is the number of individual statstics.

SPM.new_data! — Method

new_data(CH::AbstractChart)

Simulate a new observation for the control chart from the Phase 1 data, eventually modifying the underlying phase 1 object.

SPM.new_data — Method

new_data(PH2::Phase2{S,T})
new_data(PH2::Phase2{S,AbstractVector})
new_data(PH2::Phase2{S,AbstractMatrix})

Generates a new observation based on the observed Phase I (in-control) data. If it is not overloaded, then it defaults to generating data using a nonparametric bootstrap.

SPM.new_data — Method

new_data(CH::AbstractChart)

Simulate a new observation for the control chart from the Phase 1 data.

SPM.optimize_design! — Method

optimize_design!(CH, rlsim_oc[; settings = OptSettings()])

Optimizes the parameter of a simulation CH with respect to a given objective function rlsim_oc.

Arguments

CH : The simulation to optimize.

rlsim_oc : The objective function.

settings (optional, default=OptSettings()) : Optimization settings.

Returns

get_design(CH) : The optimized parameter.

SPM.optimize_design — Method

optimize_design(CH, rlsim_oc[; settings = OptSettings()])

Optimize a parameter using the specified CH and rlsim_oc.

Args

`CH`: the CH parameter.
`rlsim_oc`: the rlsim_oc parameter.
`settings`: the optimization settings.

Returns

The optimized parameter values.

SPM.optimize_grid — Method

optimize_grid(CH::ControlChart, rlconstr::Function, settings::OptSettings)

Optimizes a control chart by finding the best set of parameters using a grid search.

Args

CH: the control chart to optimize.

rlconstr: the function that evaluates the OC performance of the control chart.

settings: the optimization settings.

Returns

par_current (Vector{Float64}): the optimal set of parameters found by the optimization algorithm.

SPM.optimize_limit! — Method

optimize_limit!(CH::ControlChart[; settings = OptSettings()])

Optimizes the control limit of a ControlChart object.

Args

CH (ControlChart): The ControlChart object to optimize.

settings (OptSettings, optional): Optimization settings. Defaults to OptSettings().

Returns

The optimized control limit value.

Raises

ValueError: If the optimization method specified in settings is unknown.

Example

optimize_limit!(my_chart, settings=OptSettings(ic_solver=:SA))

SPM.optimize_limit — Method

optimize_limit(CH::ControlChart[; settings = OptSettings()])

Optimizes the control limit of a ControlChart object, without modifying the original ControlChart object.

Args

CH (ControlChart): The ControlChart object to optimize.

settings (OptSettings, optional): Optimization settings. Defaults to OptSettings().

Returns

The optimized control limit value.

Raises

ValueError: If the optimization method specified in settings is unknown.

Example

optimize_limit(my_chart, settings=OptSettings(ic_solver=:SA))

SPM.optimize_nlopt — Method

optimize_nlopt(CH::ControlChart, rlconstr::Function, settings::OptSettings)

Optimizes the Control Chart design parameter using the NLOpt library.

Args

CH: The ControlChart object to be optimized.

rlconstr: The objective function to be minimized.

settings: The settings for the optimization process. Includes:

minpar_opt: The lower bounds for the parameters.

maxpar_opt: The upper bounds for the parameters.

x_tol_opt: The relative tolerance for convergence.

maxiter_opt: The maximum number of iterations.

method_opt: The optimization method to be used.

Returns

minx::Vector{Float64}: The set of optimal parameters for the Control Chart.

SPM.run_path_sim — Method

run_path_sim(CH::AbstractChart)

Simulates a run length path for the control chart CH by sampling new data from its Phase II object.

Inputs

CH::AbstractChart - A control chart.

maxiter::Real - The maximum value of the run length. Defaults to min(maxrl(CH), 10*get_nominal_value(CH))

Returns

A vector containing values of the control chart.

SPM.run_sim — Method

run_sim(CH::AbstractChart, DGP::AbstractPhase2)

Simulates a run length for the control chart CH by sampling new data from the provided data-generating process DGP.

Inputs

CH - A control chart.

DGP - An AbstractPhase2 object.

Returns

An Int.

SPM.run_sim — Method

run_sim(CH::AbstractChart)

Simulates a run length for the control chart CH by sampling new data from its Phase II object.

Inputs

CH - A control chart.

Returns

An Int.

SPM.run_sim_oc — Method

run_sim_oc(CH::AbstractChart; shift = 0.0)

Simulates a run length under location shift for the control chart CH by sampling new data from its Phase II object.

Inputs

CH - A control chart.

shift - A location shift.

Returns

An Int.

SPM.run_sim_sa — Method

run_sim_sa(CH::AbstractChart, maxiter::Real, delta::Real)

Simulates a run length for the control chart CH by sampling new data from the Phase II object, to be used by the stochastic approximation algorithm implemented in the saCL! function.

Inputs

CH - A control chart.

maxiter - The maximum value of the run length.

delta - A value controlling how much the control limit must be shifted for the gain estimation during the first stage.

Returns

A NamedTuple containing the simulated run length, rl, the simulated run length with control limit shifted by delta, rlPlus, and the simulated run length with control limit shifted by -delta, rlMinus.

SPM.saCL! — Method

saCL!(CH::ControlChart[; rlsim::Function, settings::OptSettings])

Computes the control limit to satisfy the nominal properties of a control chart, using the stochastic approximation algorithm described in Capizzi and Masarotto (2016).

Inputs

CH - A control chart.

rlsim - A function that generates new data with signature rlsim(CH; maxiter, delta). If left unspecified, defaults to run_sim_sa. See the help for run_sim_sa for more information about the signature of the function.

settings - An OptSettings objects which contains variables that control the behaviour of the algorithm. See the Accepted settings section below for information about the settings that control the behaviour of the algorithm. For more information about the specifics of each keyword argument, see Capizzi and Masarotto (2016).

Settings

The following settings control the behaviour of the algorithm:

Nfixed_sa - The number of iterations for the gain estimation stage.

Afixed_sa - The fixed gain during the gain estimation stage.

Amin_sa - The minimum allowed value of gain.

Amax_sa - The maximum allowed value of gain.

delta_sa - The shift in control limit used during the gain estimation stage.

q_sa - The power that controls the denominator in the Robbins-Monro algorithm.

gamma_sa - The precision parameter for the stopping criterion of the algorithm.

Nmin_sa - The minimum number of iterations to avoid early terminations.

z_sa - The quantile of the Normal(0,1) that controls the probability of the stopping criterion being satisfied.

Cmrl_sa - The inflation factor for the maximum number of iterations the run length may run for.

maxiter_sa - Maximum number of iterations before the algorithm is forcibly ended.

verbose_sa - Whether to print information to the user about the state of the optimization.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Capizzi, G., & Masarotto, G. (2016). "Efficient Control Chart Calibration by Simulated Stochastic Approximation". IIE Transactions 48 (1). https://doi.org/10.1080/0740817X.2015.1055392.

SPM.saCL — Method

saCL(CH::ControlChart[; rlsim::Function, settings::OptSettings])

Applies the stochastic approximation algorithm of Capizzi and Masarotto (2016) without modifying the control chart object CH. See the documentation of saCL! for more information about the algorithm and the keyword arguments.

Returns

A NamedTuple containing the estimated control limit h, the total number of iterations iter, and information status about the convergence of the algorithm.

References

Capizzi, G., & Masarotto, G. (2016). "Efficient Control Chart Calibration by Simulated Stochastic Approximation". IIE Transactions 48 (1). https://doi.org/10.1080/0740817X.2015.1055392.

SPM.set_design! — Method

get_design(CH::AbstractChart)

Set the designs of the control chart statistic.

SPM.set_limit! — Method

function set_limit!(CH::AbstractChart, limit::AbstractLimit)

Set the control limit of a control chart.

SPM.set_nominal! — Method

set_nominal!(CH::AbstractChart, nominal::NominalProperties)

Set the nominal properties of a control chart.

SPM.set_phase1! — Method

set_phase1!(CH::AbstractChart, phase1::AbstractPhase2)

Set the Phase 1 information of a control chart.

SPM.set_statistic! — Method

function set_statistic!(CH::AbstractChart, statistic::AbstractStatistic)

Set the statistic of a control chart.

SPM.set_value! — Method

function set_value!(stat::AbstractStatistic, value)

Set the value of a statistic.

SPM.set_value! — Method

function set_value!(CH::AbstractChart, value)

Set the value of the statistic of a control chart.

SPM.shallow_copy_sim — Method

shallow_copy_sim(CH::AbstractChart)

Create a shallow copy of a control chart, so that only the statistic and the control limit are copied. This is done to prevent copying the Phase 1 data multiple times and thus reduce computational effort.

SPM.update_chart! — Method

update_chart!(CH::AbstractChart, x)

Update the control chart inplace using a new observation x.

SPM.update_chart — Method

update_chart(CH::AbstractChart, x)

Update the control chart without modifying it using a new observation x.

SPM.update_limit! — Method

update_limit!(CH::AbstractChart, x)

Update the dynamic control limit of a control chart inplace.

SPM.update_statistic — Method

update_statistic!(stat::AbstractStatistic, x)

Update a statistic with a new observation x

SPM.update_value! — Method

get_value(L::BootstrapLimit, NM::ARL)
get_value(L::BootstrapLimit, NM::QRL)

SPM.AEWMA — Type

AEWMA(λ, k, value)

Adaptive exponentially weighted moving average with design parameters λ, k, and initial value value.

The update mechanism based on a new observation x is given by

$value = (1-phi(e))*value + phi(e) * x$,

where phi(e) is a forecast error function based on the Huber function.

References

Capizzi, G. & Masarotto, G. (2003). An Adaptive Exponentially Weighted Moving Average Control Chart. Technometrics, 45(3), 199-207.

SPM.ARL — Type

ARL(value)

Value of the in-control average run length of the control chart, i.e. if $RL = \inf{t > 0: \text{Chart detects OC}}$ is the run length, then the average run length $ARL$ is

$ARL = \mathbb{E}[RL|\tau = +\infty]$,

where $\{\tau = +\infty\}$ represents the process being in-control.

SPM.BlockBootstrap — Type

BlockBootstrap{T} <: AbstractSampling

Represents a (circular) block bootstrap sampling method.

Fields

block::T: The current block of data being sampled from.

blocksize::Int: The size of each block.

t::Int: The current index within the block.

Constructors

BlockBootstrap(blocksize::Int, data::Vector{T}) where T: Constructs a BlockBootstrap object for a vector of data.

BlockBootstrap(blocksize::Int, data::Matrix{T}) where T: Constructs a BlockBootstrap object for a matrix of data.

SPM.CUSUM — Type

CUSUM(k, value, upw::Bool)

CUSUM statistic with design parameter k and initial value value.

The update mechanism based on a new observation x is given by:

if upw == true, then $value = \max\{0, value + x - k\}$;

if upw == false, then $value = \min\{0, value + x + k\}$.

References

Page, E. S. (1954). Continuous Inspection Schemes. Biometrika, 41(1/2), 100. https://doi.org/10.2307/2333009

SPM.EWMA — Type

EWMA(λ, value)

Exponentially weighted moving average with design parameter λ and initial value value.

The update mechanism based on a new observation x is given by

$value = (1-λ)*value + λ * x$.

References

Roberts, S. W. (1959). Control Chart Tests Based on Geometric Moving Averages. Technometrics, 1(3), 239-250. https://doi.org/10.1080/00401706.1959.10489860

SPM.LocationScaleStatistic — Type

LocationScaleStatistic{S, M, P}

A mutable struct representing a statistic applied to a location-scale family.

Fields

stat::S: The statistic.

μ::M: The location parameter.

Ω::P: The precision parameter (inverse of the variance).

Examples

STAT = EWMA(λ = 0.2)
RSTAT = LocationScaleStatistic(STAT, 1.0, 2.5)

SPM.MEWMA — Type

MEWMA(Λ, value)

Exponentially weighted moving average with smoothing matrix Λ and initial value value.

The update mechanism based on a new observation x is given by

$Z = (I-Λ)*Z + Λ * x$,

and the chart value is defined as

$value = Z' Λ^(-1) Z$.

References

Lowry, C. A., Woodall, W. H., Champ, C. W., & Rigdon, S. E. (1992). A Multivariate Exponentially Weighted Moving Average Control Chart. Technometrics, 34(1), 46-53. https://doi.org/10.1080/00401706.1992.10485232

SPM.OneSidedCurvedLimit — Type

OneSidedCurvedLimit(h::Float64, upw::Bool)
OneSidedCurvedLimit(h::Vector{T}, upw::Vector{Bool})

Curved one-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit.

if upw == true, $RL = \inf\{t : C_t > h\cdot f(t)\}$

if upw == false, $RL = \inf\{t : C_t < -h\cdot f(t)\}$

Note that h > 0 by the way it is defined.

SPM.OneSidedFixedLimit — Type

OneSidedFixedLimit(h::Float64, upw::Bool)

Classical fixed one-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit.

if upw == true, $RL = \inf\{t : C_t > h\}$

if upw == false, $RL = \inf\{t : C_t < h\}$

Note that h > 0 by the way it is defined.

SPM.Phase2 — Type

Phase2 is a struct that holds the reference sample data and a sampling method to generate new observations from the reference data.

Arguments

samp = Bootstrap(): The sampling method to be used to generate new observations. Defaults to Bootstrap().

data: The data obtained after phase 1.

Examples

x = randn(500) PH2 = Phase2(data = x)

SPM.Phase2Distribution — Type

Phase2Distribution{T} <: AbstractPhase2

A struct representing Phase II observations, it is used to generate and monitor new data from the true data-generating process. It contains a field dist of type T, which represents the underlying data-generating process.

Notes

A method rand(::T) is required to generate new data from dist.

Example

DGP = Phase2Distribution(Normal(0,1))
new_data(DGP)

SPM.QRL — Type

QRL(value, qtl)

Value of the in-control quantile of the run length of the control chart, i.e. if $RL = \inf{t > 0 : \text{Chart detects OC}}$ is the run length, then value is the value of the qtl-level quantile of the distribution of $RL$ if the process is in-control.

SPM.StationaryBootstrap — Type

StationaryBootstrap{T} <: AbstractSampling

Represents a stationary block bootstrap sampling method, where the block length is sampled from a Geometric random variable.

Fields

block::T: The current block of data being sampled from.

blocksize::Int: The average size of each block.

t::Int: The current index within the block.

Constructors

StationaryBootstrap(blocksize::Int, data::Vector{T}) where T: Constructs a StationaryBootstrap object for a vector of data.

StationaryBootstrap(blocksize::Int, data::Matrix{T}) where T: Constructs a StationaryBootstrap object for a matrix of data.

SPM.TwoSidedCurvedLimit — Type

TwoSidedCurvedLimit(h::Float64)
TwoSidedCurvedLimit(h::Vector{T})

Curved one-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit.

$RL = \inf\{t > 0 : |C_t| > h\cdot f(t)\}$.

Note that h > 0 by the way it is defined.

SPM.TwoSidedFixedLimit — Type

TwoSidedFixedLimit(h::Float64)

Classical fixed two-sided limit, such that the run length $RL$ of a control chart is the first time $t$ in which the statistic $C_t$ crosses the limit:

$RL = \inf\{t > 0 : |C_t| > h\}$.

Note that h > 0 by the way it is defined.